+++ lib/models/purchasereport_model.dart
@@ -1,8 +1,8 @@
 import 'package:amber_erp/models/authentication.dart';
 
-class purchasereport{
+class purchasereport {
   final String supplierName;
-  final int purchaseAmount;
+  final double purchaseAmount;
   final String paymentMode;
   final String invoiceNo;
 
@@ -11,9 +11,13 @@ class purchasereport{
     required this.purchaseAmount,
     required this.paymentMode,
     required this.invoiceNo,
-});
+  });
 }
-Future<List>getPurchasereport(DateTime today) async{
-  List<dynamic>list=(await PurchaseReportMaster().purchasereportdetail(today) as List).reversed.toList();
+
+Future<List> getPurchasereport(DateTime today) async {
+  List<dynamic> list =
+      (await PurchaseReportMaster().purchasereportdetail(today) as List)
+          .reversed
+          .toList();
   return list;
 }
