+++ lib/models/salesreport_model.dart
@@ -1,26 +1,22 @@
-
 import 'package:amber_erp/models/authentication.dart';
 
-class report{
+class report {
   final String clientName;
-  final int salesAmount;
+  final double salesAmount;
   final String paymentMode;
-
 
   const report({
     required this.clientName,
     required this.salesAmount,
     required this.paymentMode,
-});
-
+  });
 }
-Future<List> getSalesReport(DateTime today) async{
-  List<dynamic> list =( await SalesReportMaster().salesreportdetail(today) as List).reversed.toList();
+
+Future<List> getSalesReport(DateTime today) async {
+  List<dynamic> list =
+      (await SalesReportMaster().salesreportdetail(today) as List)
+          .reversed
+          .toList();
   //print(list);
   return list;
 }
-
-
-
-
-
